+++ docs/changelog.rst
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+0.0.3 (2021-11-19)
+------------------
+
+Added
+~~~~~
+
+-  Add and use :meth:`pika.decorators.halt` as the default decorator.
+
+Changed
+~~~~~~~
+
+-  Rename :meth:`pika.decorators.rescue` to :meth:`~pika.decorators.discard`.
+
 0.0.2 (2021-11-19)
 ------------------